+++ src/Entity/Tag.php
@@ -5,7 +5,13 @@ namespace App\Entity;
 use ApiPlatform\Metadata\ApiResource;
 use App\Repository\TagRepository;
 use Doctrine\ORM\Mapping as ORM;
+use Symfony\Component\Serializer\Attribute\Groups;
+use Symfony\Component\Validator\Constraints as Assert; // assertions
 
+#[ApiResource(
+    normalizationContext: ['groups' => ['tag:read']],
+    denormalizationContext: ['groups' => ['tag:write']]
+)]
 #[ORM\Entity(repositoryClass: TagRepository::class)]
 #[ApiResource]
 class Tag
@@ -16,6 +22,9 @@ class Tag
     private ?int $id = null;
 
     #[ORM\Column(length: 100)]
+    #[Groups(['tag:read', 'wish:read', 'user:read', 'write:read', 'wish:write'])]
+    #[Assert\NotBlank]
+    #[Assert\Length(max: 100)]
     private ?string $name = null;
 
     public function getId(): ?int