+++ src/Factory/CountryFactory.php
@@ -1,0 +1,52 @@
+<?php
+
+namespace App\Factory;
+
+use App\Entity\Country;
+use Zenstruck\Foundry\Persistence\PersistentProxyObjectFactory;
+
+/**
+ * @extends PersistentProxyObjectFactory<Country>
+ */
+final class CountryFactory extends PersistentProxyObjectFactory
+{
+    /**
+     * @see https://symfony.com/bundles/ZenstruckFoundryBundle/current/index.html#factories-as-services
+     *
+     * @todo inject services if required
+     */
+    public function __construct()
+    {
+    }
+
+    public static function class(): string
+    {
+        return Country::class;
+    }
+
+    /**
+     * @see https://symfony.com/bundles/ZenstruckFoundryBundle/current/index.html#model-factories
+     *
+     * @todo add your default values here
+     */
+    protected function defaults(): array|callable
+    {
+        return [
+            'code' => self::faker()->unique()->countryCode(),
+            'currency' => CurrencyFactory::new(),
+            'flag' => self::faker()->url(),
+            'name' => self::faker()->country(),
+            'updatedAt' => \DateTimeImmutable::createFromMutable(self::faker()->dateTime()),
+        ];
+    }
+
+    /**
+     * @see https://symfony.com/bundles/ZenstruckFoundryBundle/current/index.html#initialization
+     */
+    protected function initialize(): static
+    {
+        return $this
+            // ->afterInstantiate(function(Country $country): void {})
+        ;
+    }
+}