+++ src/Factory/UserFactory.php
@@ -1,0 +1,56 @@
+<?php
+
+namespace App\Factory;
+
+use App\Entity\User;
+use Zenstruck\Foundry\Persistence\PersistentProxyObjectFactory;
+
+/**
+ * @extends PersistentProxyObjectFactory<User>
+ */
+final class UserFactory extends PersistentProxyObjectFactory
+{
+    /**
+     * @see https://symfony.com/bundles/ZenstruckFoundryBundle/current/index.html#factories-as-services
+     *
+     * @todo inject services if required
+     */
+    public function __construct()
+    {
+    }
+
+    public static function class(): string
+    {
+        return User::class;
+    }
+
+    /**
+     * @see https://symfony.com/bundles/ZenstruckFoundryBundle/current/index.html#model-factories
+     *
+     * @todo add your default values here
+     */
+    protected function defaults(): array|callable
+    {
+        return [
+            'email' => self::faker()->unique()->safeEmail,
+            'password' => password_hash('123', PASSWORD_BCRYPT),
+            'nickName' => self::faker()->userName,
+            'verified' => self::faker()->boolean(80), // 80% chance że jest zweryfikowany
+            'notify' => self::faker()->boolean(50),
+            'avatar' => self::faker()->imageUrl(100, 100, 'people'),
+            'country' => CountryFactory::new(),
+            'categories' => [CategoryFactory::new()],
+            'tags' => [TagFactory::new()],
+        ];
+    }
+
+    /**
+     * @see https://symfony.com/bundles/ZenstruckFoundryBundle/current/index.html#initialization
+     */
+    protected function initialize(): static
+    {
+        return $this
+            // ->afterInstantiate(function(User $user): void {})
+        ;
+    }
+}