+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [16.02.2025]
+- ApiResource and validation config for entites with register/login DTO + Factories and Fixture for all entities
 - Entities: Category, Country, Currency, Tag, User, WishItem
 ## [15.02.2025]
 - MainController with first version of landing page and styling